requirements.txt
flask
pytest
pyyaml
flask_json
pytest-flask
invoke
requests
